+++ LostPet/CoreViews/LogInviews/Authfunc/AuthViews/AuthenticationManager.swift
