+++ LostPet/CoreViews/LogInviews/PrincipalViews/CameraView.swift
@@ -1,0 +1,20 @@
+//
+//  CameraView.swift
+//  LostPet
+//
+//  Created by Pedro Toledo on 18/8/23.
+//
+
+import SwiftUI
+
+struct CameraView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct CameraView_Previews: PreviewProvider {
+    static var previews: some View {
+        CameraView()
+    }
+}
