+++ LostPet/CoreViews/LogInviews/PrincipalViews/NavigationBar.swift
@@ -1,0 +1,20 @@
+//
+//  NavigationBar.swift
+//  LostPet
+//
+//  Created by Pedro Toledo on 18/8/23.
+//
+
+import SwiftUI
+
+struct NavigationBar: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+struct NavigationBar_Previews: PreviewProvider {
+    static var previews: some View {
+        NavigationBar()
+    }
+}
